+++ Tests/TestConsole/Vector2D.cs
@@ -1,0 +1,42 @@
+﻿using System;
+
+namespace TestConsole
+{
+    struct Vector2D
+    {
+        private double _X;
+        //private double _Y;
+
+        public double X { get { return _X; } set { _X = value; } }
+        //public double Y { get => _Y; set => _Y = value; }
+        public double Y { get; /*private*/ set; }
+
+        public double Length => Math.Sqrt(X * X + Y * Y);
+
+        public Vector2D(double X, double Y)
+        {
+            _X = X;
+            this.Y = Y;
+        }
+
+        public static Vector2D operator +(Vector2D a, Vector2D b)
+        {
+            return new Vector2D(a.X + b.X, a.Y + b.Y);
+        }
+
+        public static Vector2D operator -(Vector2D a, Vector2D b)
+        {
+            return new Vector2D(a.X - b.X, a.Y - b.Y);
+        }
+
+        public static Vector2D operator +(Vector2D a, double b)
+        {
+            return new Vector2D(a.X + b, a.Y + b);
+        }
+
+        public static Vector2D operator -(Vector2D a)
+        {
+            return new Vector2D(-a.X, -a.Y);
+        }
+    }
+}
